+++ changelog.md
@@ -31,6 +31,9 @@
     - Debounce selection mouse moving to improve performance
     - Fixed updating split view when scrolling
     - Fixed split view selection performance issues
+    - Fixed scroll position on resize
+    - Fixed scrolling
+  - Fixed scroll lock scrolling to end when not in split screen mode
   - ANSI Parser: Fixed extra characters being added after : is detected, fixes raw ansi logging
 
 ## v0.6.4 2018-10-22